Fix bug with crash when doing large allocations
While trying to run a chroma job with a very large local volume I got a
crash due to allocation failure. I tracked this down and the error comes
from integer overflow. Using a normal int instead of size_t for holding
the size of an allocation is not the best idea. This commit changes
that.